# tools/ftd-level-editor/openapi.json
{
  "components": {
    "schemas": {
...
      "CandidateResponse": {
        "additionalProperties": false,
        "properties": {
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "candidateId": {
            "title": "Candidateid",
            "type": "string"
          },
          "catalogRevision": {
            "title": "Catalogrevision",
            "type": "string"
          },
          "changelog": {
            "title": "Changelog",
            "type": "string"
          },
          "digest": {
            "title": "Digest",
            "type": "string"
          },
          "levelIds": {
            "items": {
              "type": "string"
            },
            "title": "Levelids",
            "type": "array"
          },
          "sequenceVersion": {
            "title": "Sequenceversion",
            "type": "string"
          },
          "sourceRevision": {
            "title": "Sourcerevision",
            "type": "string"
          }
        },
        "required": [
          "candidateId",
          "sequenceVersion",
          "levelIds",
          "catalogRevision",
          "changelog",
          "actor",
          "sourceRevision",
          "digest"
        ],
        "title": "CandidateResponse",
        "type": "object"
      },
...
      "ExportDryRunRequest": {
        "additionalProperties": false,
        "properties": {
          "revision": {
            "title": "Revision",
            "type": "string"
          },
          "sessionId": {
            "title": "Sessionid",
            "type": "string"
          }
        },
        "required": [
          "sessionId",
          "revision"
        ],
        "title": "ExportDryRunRequest",
        "type": "object"
      },
      "ExportDryRunResponse": {
        "additionalProperties": false,
        "properties": {
          "dogCount": {
            "title": "Dogcount",
            "type": "integer"
          },
          "levelId": {
            "title": "Levelid",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "string"
          },
          "sessionId": {
            "title": "Sessionid",
            "type": "string"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "sessionId",
          "revision",
          "valid",
          "levelId",
          "dogCount"
        ],
        "title": "ExportDryRunResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PrepareSequenceRequest": {
        "additionalProperties": false,
        "properties": {
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "catalogRevision": {
            "title": "Catalogrevision",
            "type": "string"
          },
          "changelog": {
            "title": "Changelog",
            "type": "string"
          },
          "levelIds": {
            "items": {
              "type": "string"
            },
            "title": "Levelids",
            "type": "array"
          },
          "sequenceVersion": {
            "title": "Sequenceversion",
            "type": "string"
          },
          "sourceRevision": {
            "title": "Sourcerevision",
            "type": "string"
          }
        },
        "required": [
          "sequenceVersion",
          "levelIds",
          "catalogRevision",
          "changelog",
          "actor",
          "sourceRevision"
        ],
        "title": "PrepareSequenceRequest",
        "type": "object"
      },
      "ProtectedSequenceRequest": {
        "additionalProperties": false,
        "properties": {
          "candidateId": {
            "title": "Candidateid",
            "type": "string"
          },
          "grantId": {
            "title": "Grantid",
            "type": "string"
          },
          "remote": {
            "default": true,
            "title": "Remote",
            "type": "boolean"
          }
        },
        "required": [
          "candidateId",
          "grantId"
        ],
        "title": "ProtectedSequenceRequest",
        "type": "object"
      },
      "PublishingSnapshotResponse": {
        "additionalProperties": false,
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/CandidateResponse"
            },
            "title": "Candidates",
            "type": "array"
          },
          "remoteEnabled": {
            "title": "Remoteenabled",
            "type": "boolean"
          },
          "rollbackEligibleCandidateIds": {
            "items": {
              "type": "string"
            },
            "title": "Rollbackeligiblecandidateids",
            "type": "array"
          },
          "sagas": {
            "items": {
              "$ref": "#/components/schemas/SagaResponse"
            },
            "title": "Sagas",
            "type": "array"
          },
          "selected": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CandidateResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "selected",
          "candidates",
          "sagas",
          "rollbackEligibleCandidateIds",
          "remoteEnabled"
        ],
        "title": "PublishingSnapshotResponse",
        "type": "object"
      },
...
      "SagaResponse": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "enum": [
              "publish",
              "rollback"
            ],
            "title": "Action",
            "type": "string"
          },
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "candidateId": {
            "title": "Candidateid",
            "type": "string"
          },
          "changelog": {
            "title": "Changelog",
            "type": "string"
          },
          "digest": {
            "title": "Digest",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "remote": {
            "title": "Remote",
            "type": "boolean"
          },
          "sagaId": {
            "title": "Sagaid",
            "type": "string"
          },
          "sourceRevision": {
            "title": "Sourcerevision",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending_remote",
              "reconciling",
              "remote_committed",
              "finalizing",
              "succeeded",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "sagaId",
          "action",
          "candidateId",
          "digest",
          "actor",
          "changelog",
          "sourceRevision",
          "status",
          "remote",
          "error"
        ],
        "title": "SagaResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "LaunchCredential": {
        "in": "header",
        "name": "X-FTD-Launch-Credential",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Find the Dog Level Editor",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/publishing": {
      "get": {
        "operationId": "getPublishingSnapshot",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishingSnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "LaunchCredential": []
          }
        ],
        "summary": "Get Snapshot",
        "x-ftd-authorization": "launch-credential",
        "x-ftd-cost": "none",
        "x-ftd-side-effects": "none"
      }
    },
    "/api/publishing/activate": {
      "post": {
        "operationId": "activateSequencePublication",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProtectedSequenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SagaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "LaunchCredential": []
          }
        ],
        "summary": "Activate",
        "x-ftd-approval": "single-use-digest-bound",
        "x-ftd-cost": "none",
        "x-ftd-reconciliation": "exact-remote-readback-no-blind-republish",
        "x-ftd-revision": "digest-and-base-bound",
        "x-ftd-side-effects": "public-sequence-selection"
      }
    },
    "/api/publishing/export-dry-run": {
      "post": {
        "operationId": "dryRunCurrentSessionExport",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportDryRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportDryRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "LaunchCredential": []
          }
        ],
        "summary": "Export Dry Run",
        "x-ftd-artifacts": "validation-facts-only",
        "x-ftd-cost": "none",
        "x-ftd-revision": "bound",
        "x-ftd-side-effects": "none"
      }
    },
    "/api/publishing/previews": {
      "post": {
        "operationId": "prepareSequencePublication",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrepareSequenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "LaunchCredential": []
          }
        ],
        "summary": "Prepare",
        "x-ftd-approval": "not-required",
        "x-ftd-cost": "none",
        "x-ftd-revision": "catalog-bound",
        "x-ftd-side-effects": "immutable-local-preview"
      }
    },
    "/api/publishing/rollback": {
      "post": {
        "operationId": "rollbackSequencePublication",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProtectedSequenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SagaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "LaunchCredential": []
          }
        ],
        "summary": "Rollback",
        "x-ftd-approval": "single-use-digest-bound",
        "x-ftd-cost": "none",
        "x-ftd-reconciliation": "exact-remote-readback-no-blind-republish",
        "x-ftd-revision": "digest-and-base-bound",
        "x-ftd-side-effects": "public-sequence-selection"
      }
    },
    "/api/publishing/sagas/{saga_id}/reconcile": {
      "post": {
        "operationId": "reconcileSequencePublication",
        "parameters": [
          {
            "in": "path",
            "name": "saga_id",
            "required": true,
            "schema": {
              "title": "Saga Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SagaResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "LaunchCredential": []
          }
        ],
        "summary": "Reconcile",
        "x-ftd-cost": "none",
        "x-ftd-reconciliation": "readback-only-no-publish",
        "x-ftd-side-effects": "remote-readback-and-local-finalization"
      }
    },
...
  }
}